Generic addon-events decorator
